refactor(project): fix imports
fixes #87
